refactor: rename onchain/offchain to use protocol name instead
