+++ apps/ui/src/components/SpaceCreateSnapshot.vue
